rename cmds to import and export.
