+++ src/context/yaml/index.js
@@ -31,11 +31,13 @@ export default class {
 
   loadFile(f) {
     let toLoad = path.join(this.basePath, f);
+    const p = path;
+    const a = fs;
     if (!isFile(toLoad)) {
       // try load not relative to yaml file
       toLoad = f;
     }
-    return loadFile(toLoad, this.mappings);
+    return loadFile(path.resolve(toLoad), this.mappings);
   }
 
   async load() {
@@ -84,17 +86,17 @@ export default class {
 
     await Promise.all(Object.entries(handlers).map(async ([ name, handler ]) => {
       try {
-        const dumped = await handler.dump(this);
-        if (dumped) {
-          log.info(`Dumping ${name}`);
-          Object.entries(dumped)
+        const data = await handler.dump(this);
+        if (data) {
+          log.info(`Exporting ${name}`);
+          Object.entries(data)
             .forEach(([ k, v ]) => {
               this.assets[k] = v;
             });
         }
       } catch (err) {
         log.error(err.stack);
-        throw new Error(`Problem dumping ${name}`);
+        throw new Error(`Problem exporting ${name}`);
       }
     }));
